requirements.txt
appdirs==1.4.4
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cycler==0.10.0
decorator==4.4.2

ffmpeg-python==0.2.0
Flask==1.1.2
future==0.18.2
idna==2.9
imagecodecs==2020.2.18
imageio==2.8.0
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
networkx==2.4
numpy==1.18.4
opencv-python==4.2.0.34
packaging==20.3
Pillow==7.1.2
pooch==1.1.0
pydub==0.24.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
requests==2.23.0
scikit-image==0.17.1
scikit-video==1.1.11
scipy==1.4.1
six==1.14.0
tifffile==2020.5.11


tqdm==4.46.0
urllib3==1.25.9
Werkzeug==1.0.1
